feat: refactor the guards
