+++ src/app/app-routing.module.ts
@@ -23,9 +23,9 @@ import { ResendConfirmEmailComponent } from './view/resend-confirmation-email/re
 import { DashboardComponent } from './view/dashboard/dashboard.component';
 import { GuestGuard } from './guards/guest.guard';
 import { AuthUnverifiedGuard } from './guards/auth-unverified.guard';
-import { EmailVerificationGuard } from './guards/email-verified.guard';
-import { TermsVerificationGuard } from './guards/terms-verified.guard';
-import { PhoneVerificationGuard } from './guards/phone-verification.guard';
+import { EmailVerificationGuard } from './guards/verification/email-verified.guard';
+import { TermsVerificationGuard } from './guards/verification/terms-verified.guard';
+import { PhoneVerificationGuard } from './guards/verification/phone-verification.guard';
 
 ('./view/UserAcceptInvitationComponent');
 const routes: Routes = [
