+++ src/app/guards/verification/email-verified.guard.ts
@@ -1,14 +1,13 @@
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthGuard } from './auth.guard';
-import { UserService } from '../auth/services/user.service';
+import { AuthGuard } from '../auth.guard';
+import { UserService } from '../../auth/services/user.service';
 import { inject } from '@angular/core';
 import { map } from 'rxjs/operators';
 
-export const OrganizationDocumentsGuard: CanActivateFn = (route, state) => {
+export const EmailVerificationGuard: CanActivateFn = (route, state) => {
   const userService = inject(UserService);
   const router = inject(Router);
 
-  // First check if user is logged in
   const loggedInResult = AuthGuard(route, state);
   if (loggedInResult !== true) {
     return loggedInResult;
@@ -16,7 +15,7 @@ export const OrganizationDocumentsGuard: CanActivateFn = (route, state) => {
 
   return userService.userProfile().pipe(
     map((user) => {
-      if (user.organization?.verifiedAt) {
+      if (user.emailVerifiedAt) {
         return router.parseUrl('/dashboard');
       }
