+++ src/app/guards/verification/terms-verified.guard.ts
@@ -1,16 +1,13 @@
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthGuard } from './auth.guard';
-import { UserService } from '../auth/services/user.service';
+import { AuthGuard } from '../auth.guard';
+import { UserService } from '../../auth/services/user.service';
 import { inject } from '@angular/core';
 import { map } from 'rxjs/operators';
 
-export const EmailVerificationGuard: CanActivateFn = (route, state) => {
+export const TermsVerificationGuard: CanActivateFn = (route, state) => {
   const userService = inject(UserService);
   const router = inject(Router);
 
-  console.log('EmailVerificationGuard activated');
-
-  // First check if user is logged in
   const loggedInResult = AuthGuard(route, state);
   if (loggedInResult !== true) {
     return loggedInResult;
@@ -18,7 +15,7 @@ export const EmailVerificationGuard: CanActivateFn = (route, state) => {
 
   return userService.userProfile().pipe(
     map((user) => {
-      if (user.emailVerifiedAt) {
+      if (user.termsAcceptedAt) {
         return router.parseUrl('/dashboard');
       }
